Fix mech pilot status immunity inconsistency
In the base game (and Arcana), the player is immune from status effects
being applied while in a mech. However, if a persistent status effect is
already applied to a player, it continues its effect after the player
enters a mech. If the effect causes damage, the player will continue to
take damage while in the mech. This patch fixes this bug for Arcana's
environment hazards by ensuring that they consistently do not apply to
players inside mechs. (The fix for the base game's environment hazards
is provided by the Better Environment Hazards mod, but it is preferable
to apply the fix directly to mods that provide their own environment
hazards.)